requirements.txt
aiofiles==0.3.1
aiohttp==2.3.1
aiosqlite==0.2.0
beautifulsoup4==4.6.0
django==1.11.6
djangorestframework==3.7.1
pillow==4.3.0
tensorflow==1.3.0
